requirements.txt
streamlit==1.41.1
pandas==2.2.0
mysql-connector-python==8.2.0
plotly==5.24.1
flask==3.0.0
axios==0.21.1